requirements.txt
aniso8601==4.0.1
certifi==2018.10.15
chardet==3.0.4
Click==7.0
Flask==1.0.2
Flask-RESTful==0.3.6
gunicorn==19.9.0
idna==2.7
influxdb==5.2.0
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
paho-mqtt==1.4.0
python-dateutil==2.7.5
pytz==2018.7
PyYAML==3.13
requests==2.20.1
six==1.11.0
urllib3==1.24.1
Werkzeug==0.14.1
